test: added framework and test for wallet api/live app detox test
