+++ apps/ledger-live-desktop/tests/utils/dummy-wallet-app/src/react-app-env.d.ts
@@ -8,4 +8,7 @@ interface Window {
       };
     };
   };
+  ReactNativeWebview?: {
+    postMessage: (value: unknown) => void;
+  };
 }
