+++ apps/ledger-live-mobile/e2e/models/liveApps/cryptoDrawer.ts
@@ -1,0 +1,11 @@
+import { tapByText } from "../../helpers";
+
+export default class CryptoDrawer {
+  selectCurrencyFromDrawer(currencyName: string) {
+    return tapByText(currencyName);
+  }
+
+  selectAccountFromDrawer(accountName: string) {
+    return tapByText(accountName);
+  }
+}
